scripts/requirements.txt
osmnx>=1.9.0
geopandas>=0.14.0
psycopg2-binary>=2.9.0
shapely>=2.0.0
python-dotenv>=1.0.0
